backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
openai==1.6.1
langchain-groq
